requirements.txt
pyrogram
tgcrypto
aiohttp
redis
python-dotenv